+++ src/main/java/io/xlorey/fluxloader/utils/Constants.java
@@ -32,6 +32,11 @@ public class Constants {
         public static final String PLUGINS_DEFAULT_CONFIG_NAME = "config.yml";
 
         /**
+         * Plugin metadata file name
+         */
+        public static final String PLUGINS_METADATA_NAME = "metadata.json";
+
+        /**
          * Path to the FluxLoader logo
          */
         public static final String PATH_FLUXLOADER_LOGO = "io/xlorey/fluxloader/media/FluxLoader-Logo.png";
@@ -49,7 +54,7 @@ public class Constants {
         /**
          * Link to project github
          */
-        public static final String GITHUB_LINK = "https://github.com/xLorey/FluxLoader-PZ";
+        public static final String GITHUB_LINK = "https://github.com/xLorey/FluxLoader";
 
         /**
          * Link to the project's discord channel